Improved the usability of move command

+++ src/snake_pluggin.rs
@@ -218,7 +218,7 @@ pub fn spawn_snake_system(
                 .with_children(|parent| {
                     parent.spawn(SnakePartSpriteBundle::new(
                         Vec2::ONE,
-                        SNAKE_COLORS[snake_index as usize],
+                        SNAKE_COLORS[snake_index],
                     ));
                 });
         }
@@ -260,7 +260,7 @@ pub fn select_snake_mouse_system(
 
     let (camera, camera_transform) = camera.single();
     let mouse_world_position = {
-        let window_size = Vec2::new(window.width() as f32, window.height() as f32);
+        let window_size = Vec2::new(window.width(), window.height());
         let ndc = (mouse_position / window_size) * 2.0 - Vec2::ONE;
         let ndc_to_world = camera_transform.compute_matrix() * camera.projection_matrix().inverse();
         let world_pos = ndc_to_world.project_point3(ndc.extend(-1.0));
